requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
tensorflow==2.16.1
python-dotenv==1.0.1
supabase==2.3.3
gotrue==2.1.0
rich==13.7.1
streamlit-aggrid
